Update docs, many refactors and fixes
