+++ griptape/drivers/structure_run/base_structure_run_driver.py
@@ -1,11 +1,18 @@
 from abc import ABC, abstractmethod
+
 from attrs import define
 
-from griptape.artifacts.base_artifact import BaseArtifact
+from griptape.artifacts import BaseArtifact, ErrorArtifact
 
 
 @define
 class BaseStructureRunDriver(ABC):
+    def run(self, *args: BaseArtifact) -> BaseArtifact:
+        try:
+            return self.try_run(*args)
+        except Exception as e:
+            return ErrorArtifact(str(e))
+
     @abstractmethod
-    def run(self, *args) -> BaseArtifact:
+    def try_run(self, *args: BaseArtifact) -> BaseArtifact:
         ...
